+++ tests/read_exact.rs
@@ -12,9 +12,7 @@ use std::time::Duration;
 use tokio::io::{AsyncReadExt, AsyncWriteExt};
 use tokio::time::sleep;
 
-use shared_files::{
-    FileSize, SharedFile, SharedTemporaryFile, SharedTemporaryFileReader,
-};
+use shared_files::{FileSize, SharedFile, SharedTemporaryFile, SharedTemporaryFileReader};
 
 /// The number of u16 values to prefill the file with.
 const NUM_PREFILL_VALUES_U16: usize = 65_536;
@@ -75,9 +73,7 @@ async fn read_exact() {
 fn validate_result(read: Vec<u8>) {
     assert_eq!(read.len(), NUM_BYTES);
     read.chunks_exact(2)
-        .into_iter()
         .map(|a| u16::from_ne_bytes([a[0], a[1]]))
-        .into_iter()
         .enumerate()
         .for_each(|(i, value)| assert_eq!(value, i as u16));
 }
@@ -88,10 +84,7 @@ async fn prefill_file(file: SharedTemporaryFile) -> SharedTemporaryFile {
     {
         let mut writer = file.writer().await.expect("failed to create writer");
         for _ in 0..NUM_PREFILL_VALUES_U16 {
-            writer
-                .write_u16_le(0 as u16)
-                .await
-                .expect("failed to write");
+            writer.write_u16_le(0_u16).await.expect("failed to write");
         }
 
         // Ensure data is flushed to disk.